+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 [![Donate via PayPal](https://img.shields.io/badge/Donate-PayPal-blue.svg?style=flat-square)](https://www.paypal.me/techtoday) 
 
 <p>
+<b>Version 2.0.1</b> August 2023<br/>
+- NEW: added Elevenlabs TTS engine https://elevenlabs.io.<br/>
+</p>
+<p>
 <b>Version 2.0.0</b> June 2023<br/>
 - Bumped paid Google TTS, free google TTS and Ms TTS dependencies.<br/>
 - Moved help to the node-red's help pane.<br/>